Add Exceptions Directory
